+++ lib/src/material_controls.dart
@@ -1,9 +1,9 @@
 import 'dart:async';
 
-import 'package:chewie/src/chewie_player.dart';
-import 'package:chewie/src/chewie_progress_colors.dart';
-import 'package:chewie/src/material_progress_bar.dart';
-import 'package:chewie/src/utils.dart';
+import 'package:chewie_audio/src/chewie_player.dart';
+import 'package:chewie_audio/src/chewie_progress_colors.dart';
+import 'package:chewie_audio/src/material_progress_bar.dart';
+import 'package:chewie_audio/src/utils.dart';
 import 'package:flutter/material.dart';
 import 'package:video_player/video_player.dart';
 
@@ -19,18 +19,15 @@ class MaterialControls extends StatefulWidget {
 class _MaterialControlsState extends State<MaterialControls> {
   VideoPlayerValue _latestValue;
   double _latestVolume;
-  bool _hideStuff = true;
-  Timer _hideTimer;
   Timer _initTimer;
   Timer _showAfterExpandCollapseTimer;
   bool _dragging = false;
-  bool _displayTapped = false;
 
   final barHeight = 48.0;
   final marginSize = 5.0;
 
   VideoPlayerController controller;
-  ChewieController chewieController;
+  ChewieAudioController chewieController;
 
   @override
   Widget build(BuildContext context) {
@@ -49,32 +46,7 @@ class _MaterialControlsState extends State<MaterialControls> {
             );
     }
 
-    return MouseRegion(
-      onHover: (_) {
-        _cancelAndRestartTimer();
-      },
-      child: GestureDetector(
-        onTap: () => _cancelAndRestartTimer(),
-        child: AbsorbPointer(
-          absorbing: _hideStuff,
-          child: Column(
-            children: <Widget>[
-              _latestValue != null &&
-                          !_latestValue.isPlaying &&
-                          _latestValue.duration == null ||
-                      _latestValue.isBuffering
-                  ? const Expanded(
-                      child: const Center(
-                        child: const CircularProgressIndicator(),
-                      ),
-                    )
-                  : _buildHitArea(),
-              _buildBottomBar(context),
-            ],
-          ),
-        ),
-      ),
-    );
+    return _buildBottomBar(context);
   }
 
   @override
@@ -85,7 +57,6 @@ class _MaterialControlsState extends State<MaterialControls> {
 
   void _dispose() {
     controller.removeListener(_updateState);
-    _hideTimer?.cancel();
     _initTimer?.cancel();
     _showAfterExpandCollapseTimer?.cancel();
   }
@@ -93,7 +64,7 @@ class _MaterialControlsState extends State<MaterialControls> {
   @override
   void didChangeDependencies() {
     final _oldController = chewieController;
-    chewieController = ChewieController.of(context);
+    chewieController = ChewieAudioController.of(context);
     controller = chewieController.videoPlayerController;
 
     if (_oldController != chewieController) {
@@ -104,104 +75,25 @@ class _MaterialControlsState extends State<MaterialControls> {
     super.didChangeDependencies();
   }
 
-  AnimatedOpacity _buildBottomBar(
+  Widget _buildBottomBar(
     BuildContext context,
   ) {
     final iconColor = Theme.of(context).textTheme.button.color;
 
-    return AnimatedOpacity(
-      opacity: _hideStuff ? 0.0 : 1.0,
-      duration: Duration(milliseconds: 300),
-      child: Container(
-        height: barHeight,
-        color: Theme.of(context).dialogBackgroundColor,
-        child: Row(
-          children: <Widget>[
-            _buildPlayPause(controller),
-            chewieController.isLive
-                ? Expanded(child: const Text('LIVE'))
-                : _buildPosition(iconColor),
-            chewieController.isLive ? const SizedBox() : _buildProgressBar(),
-            chewieController.allowMuting
-                ? _buildMuteButton(controller)
-                : Container(),
-            chewieController.allowFullScreen
-                ? _buildExpandButton()
-                : Container(),
-          ],
-        ),
-      ),
-    );
-  }
-
-  GestureDetector _buildExpandButton() {
-    return GestureDetector(
-      onTap: _onExpandCollapse,
-      child: AnimatedOpacity(
-        opacity: _hideStuff ? 0.0 : 1.0,
-        duration: Duration(milliseconds: 300),
-        child: Container(
-          height: barHeight,
-          margin: EdgeInsets.only(right: 12.0),
-          padding: EdgeInsets.only(
-            left: 8.0,
-            right: 8.0,
-          ),
-          child: Center(
-            child: Icon(
-              chewieController.isFullScreen
-                  ? Icons.fullscreen_exit
-                  : Icons.fullscreen,
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-
-  Expanded _buildHitArea() {
-    return Expanded(
-      child: GestureDetector(
-        onTap: () {
-          if (_latestValue != null && _latestValue.isPlaying) {
-            if (_displayTapped) {
-              setState(() {
-                _hideStuff = true;
-              });
-            } else
-              _cancelAndRestartTimer();
-          } else {
-            _playPause();
-
-            setState(() {
-              _hideStuff = true;
-            });
-          }
-        },
-        child: Container(
-          color: Colors.transparent,
-          child: Center(
-            child: AnimatedOpacity(
-              opacity:
-                  _latestValue != null && !_latestValue.isPlaying && !_dragging
-                      ? 1.0
-                      : 0.0,
-              duration: Duration(milliseconds: 300),
-              child: GestureDetector(
-                child: Container(
-                  decoration: BoxDecoration(
-                    color: Theme.of(context).dialogBackgroundColor,
-                    borderRadius: BorderRadius.circular(48.0),
-                  ),
-                  child: Padding(
-                    padding: EdgeInsets.all(12.0),
-                    child: Icon(Icons.play_arrow, size: 32.0),
-                  ),
-                ),
-              ),
-            ),
-          ),
-        ),
+    return Container(
+      height: barHeight,
+      color: Theme.of(context).dialogBackgroundColor,
+      child: Row(
+        children: <Widget>[
+          _buildPlayPause(controller),
+          chewieController.isLive
+              ? Expanded(child: const Text('LIVE'))
+              : _buildPosition(iconColor),
+          chewieController.isLive ? const SizedBox() : _buildProgressBar(),
+          chewieController.allowMuting
+              ? _buildMuteButton(controller)
+              : Container(),
+        ],
       ),
     );
   }
@@ -211,8 +103,6 @@ class _MaterialControlsState extends State<MaterialControls> {
   ) {
     return GestureDetector(
       onTap: () {
-        _cancelAndRestartTimer();
-
         if (_latestValue.volume == 0) {
           controller.setVolume(_latestVolume ?? 0.5);
         } else {
@@ -220,22 +110,18 @@ class _MaterialControlsState extends State<MaterialControls> {
           controller.setVolume(0.0);
         }
       },
-      child: AnimatedOpacity(
-        opacity: _hideStuff ? 0.0 : 1.0,
-        duration: Duration(milliseconds: 300),
-        child: ClipRect(
+      child: ClipRect(
+        child: Container(
           child: Container(
-            child: Container(
-              height: barHeight,
-              padding: EdgeInsets.only(
-                left: 8.0,
-                right: 8.0,
-              ),
-              child: Icon(
-                (_latestValue != null && _latestValue.volume > 0)
-                    ? Icons.volume_up
-                    : Icons.volume_off,
-              ),
+            height: barHeight,
+            padding: EdgeInsets.only(
+              left: 8.0,
+              right: 8.0,
+            ),
+            child: Icon(
+              (_latestValue != null && _latestValue.volume > 0)
+                  ? Icons.volume_up
+                  : Icons.volume_off,
             ),
           ),
         ),
@@ -280,46 +166,10 @@ class _MaterialControlsState extends State<MaterialControls> {
     );
   }
 
-  void _cancelAndRestartTimer() {
-    _hideTimer?.cancel();
-    _startHideTimer();
-
-    setState(() {
-      _hideStuff = false;
-      _displayTapped = true;
-    });
-  }
-
   Future<Null> _initialize() async {
     controller.addListener(_updateState);
 
     _updateState();
-
-    if ((controller.value != null && controller.value.isPlaying) ||
-        chewieController.autoPlay) {
-      _startHideTimer();
-    }
-
-    if (chewieController.showControlsOnInitialize) {
-      _initTimer = Timer(Duration(milliseconds: 200), () {
-        setState(() {
-          _hideStuff = false;
-        });
-      });
-    }
-  }
-
-  void _onExpandCollapse() {
-    setState(() {
-      _hideStuff = true;
-
-      chewieController.toggleFullScreen();
-      _showAfterExpandCollapseTimer = Timer(Duration(milliseconds: 300), () {
-        setState(() {
-          _cancelAndRestartTimer();
-        });
-      });
-    });
   }
 
   void _playPause() {
@@ -327,12 +177,8 @@ class _MaterialControlsState extends State<MaterialControls> {
 
     setState(() {
       if (controller.value.isPlaying) {
-        _hideStuff = false;
-        _hideTimer?.cancel();
         controller.pause();
       } else {
-        _cancelAndRestartTimer();
-
         if (!controller.value.initialized) {
           controller.initialize().then((_) {
             controller.play();
@@ -344,14 +190,6 @@ class _MaterialControlsState extends State<MaterialControls> {
           controller.play();
         }
       }
-    });
-  }
-
-  void _startHideTimer() {
-    _hideTimer = Timer(const Duration(seconds: 3), () {
-      setState(() {
-        _hideStuff = true;
-      });
     });
   }
 
@@ -371,15 +209,11 @@ class _MaterialControlsState extends State<MaterialControls> {
             setState(() {
               _dragging = true;
             });
-
-            _hideTimer?.cancel();
           },
           onDragEnd: () {
             setState(() {
               _dragging = false;
             });
-
-            _startHideTimer();
           },
           colors: chewieController.materialProgressColors ??
               ChewieProgressColors(